feat(scheduled-tasks): move runs to a right-side panel tab (scheduled chats only), remove the left-sidebar navigator
